requirements.txt
numpy
pywin32
PyQt5
pyautogui
Pillow